+++ tests/feature/test_report.py
@@ -120,6 +120,7 @@ def test_step_trace(pytester):
         "keyword": "Scenario",
         "line_number": 5,
         "name": "Passing",
+        "description": "",
         "steps": [
             {
                 "duration": OfType(float),
@@ -159,6 +160,7 @@ def test_step_trace(pytester):
         "keyword": "Scenario",
         "line_number": 10,
         "name": "Failing",
+        "description": "",
         "steps": [
             {
                 "duration": OfType(float),
@@ -197,6 +199,7 @@ def test_step_trace(pytester):
         "keyword": "Scenario Outline",
         "line_number": 14,
         "name": "Outlined",
+        "description": "",
         "steps": [
             {
                 "duration": OfType(float),
@@ -243,6 +246,7 @@ def test_step_trace(pytester):
         "keyword": "Scenario Outline",
         "line_number": 14,
         "name": "Outlined",
+        "description": "",
         "steps": [
             {
                 "duration": OfType(float),